+++ apps/demo-react/util/walletconnectProjectId.ts
@@ -1,0 +1,1 @@
+export const WC_PROJECT_ID = 'cbbbf9cd4c2a5581edd36dc8cabe664f';