+++ packages/connect-wallet-modal/src/components/WalletsModal/types.ts
@@ -19,7 +19,6 @@ export type WalletsModalProps = ModalProps & {
   shouldInvertWalletIcon?: boolean;
   buttonsFullWidth?: boolean;
   metrics?: Metrics;
-  walletConnectProjectId?: string;
 };
 
 export type ButtonsCommonProps = {
@@ -29,5 +28,4 @@ export type ButtonsCommonProps = {
   shouldInvertWalletIcon: boolean;
   setRequirements(isVisible: boolean, requirementsData: RequirementsData): void;
   metrics?: Metrics;
-  walletConnectProjectId?: string;
 };