+++ packages/core-react/src/hooks/useReefKnotContext.ts
@@ -1,6 +1,5 @@
 import { useContext } from 'react';
 import { ReefKnotContext, ReefKnotContextValue } from '../context';
 
-export const useReefKnotContext = (): ReefKnotContextValue => {
-  return useContext(ReefKnotContext);
-};
+export const useReefKnotContext = (): ReefKnotContextValue =>
+  useContext(ReefKnotContext);