+++ packages/wallets-helpers/src/factories/walletConnectConnector.ts
@@ -1,4 +1,5 @@
 import { WalletConnectLegacyConnector } from 'wagmi/connectors/walletConnectLegacy';
+import { WalletConnectConnector } from 'wagmi/connectors/walletConnect';
 import { isValidHttpUrl } from '../utils';
 import { walletConnectMobileLinks } from './walletConnectMobileLinks';
 
@@ -16,13 +17,35 @@ export const prepareWalletConnectRPC = (rpc: Record<number, string>) => {
 
 export const getWalletConnectConnector = ({
   rpc,
-  noMobileLinks,
-  qrcode,
+  projectId = '',
+  noMobileLinks = false,
+  qrcode = true,
+  v2: _v2 = false,
 }: {
   rpc: Record<number, string>;
+  projectId?: string;
   noMobileLinks?: boolean;
   qrcode?: boolean;
+  v2?: boolean;
 }) => {
+  // WalletConnect v2 will automatically replace legacy v1 after this date:
+  const v2TransitionDate = new Date('2023-06-20T00:00:00Z');
+  const v2 = _v2 || new Date() > v2TransitionDate;
+  if (v2) {
+    return new WalletConnectConnector({
+      options: {
+        projectId,
+        showQrModal: qrcode,
+        qrModalOptions: {
+          explorerAllowList: undefined,
+          explorerDenyList: undefined,
+          themeVariables: {
+            '--w3m-z-index': '1000',
+          },
+        },
+      },
+    });
+  }
   return new WalletConnectLegacyConnector({
     options: {
       rpc: prepareWalletConnectRPC(rpc),