+++ packages/web3-react/src/hooks/useConnectors.ts
@@ -1,5 +1,8 @@
 import { useContext } from 'react';
-import { ConnectorsContext, ConnectorsContextValue } from '../context';
+import {
+  ConnectorsContext,
+  ConnectorsContextValue,
+} from '../context/connectors';
 
 export const useConnectors = (): ConnectorsContextValue => {
   return useContext(ConnectorsContext);