+++ packages/wallets/ambire/CHANGELOG.md
@@ -1,9 +1,17 @@
 # @reef-knot/wallet-adapter-ambire
 
-## 2.0.0
+## 1.1.1
 
 ### Patch Changes
 
-- Updated dependencies [faa8815]
+- Pass walletconnectProjectId to be ready for WC v2
+- Updated dependencies
+  - @reef-knot/types@1.1.1
+  - @reef-knot/wallets-helpers@1.0.1
+
+## 1.1.0
+
+### Patch Changes
+
 - Updated dependencies [faa8815]
   - @reef-knot/types@1.1.0